Create IResponseWriter and made ResponseWriterWrapper implement it. Also moved dto/, repository/ and util/ to internal/
