+++ internal/dto/abstractRequest.go
