+++ internal/dto/sendNotificationRequest.go
