+++ internal/dto/templateRecord.go
