+++ internal/util/iResponseWriter.go
@@ -1,0 +1,8 @@
+package util
+
+type IResponseWriter[T any] interface {
+	Status(statusCode int) (*T)
+	Text(text string) (*T)
+	Json(data interface{}) (*T)
+	Bytes(data []byte) (*T)
+}
